+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/internal/ui/editors/schematic/border/BorderUtil.java
@@ -67,7 +67,7 @@ public class BorderUtil
 			i = x;
 		}
 
-		return (int) ( y * i / x );
+		return  y * i / x ;
 	}
 
 	/**
